Fixes to compile on windows with new Neuron annotator code:   Avoid including windows.h too late, for the sad cases where it must be included.   Disambiguate some std::sqrt calls.   Add return values to methods that lacked them.

+++ v3d_main/neuron_annotator/CellCounter3D.cpp
@@ -288,7 +288,7 @@ void CellCounter3D::markImage() {
                     } else if (ex>=xDim) {
                         sx=xDim-1;
                     }
-                    double distance = std::sqrt((sz-z)*(sz-z)+(sy-y)*(sy-y)+(sx-x)*(sx-x));
+                    double distance = std::sqrt((double)((sz-z)*(sz-z)+(sy-y)*(sy-y)+(sx-x)*(sx-x)));
                     if (distance<=MARK_RADIUS) {
                         unsigned char signal=data[0][sz][sy][sx];
                         int r=(MARK_COLOR[0]*signal)/255;
